requirements.txt
requests
pytesseract
pillow
redis
bs4
toolkity>=1.1.3
